==> notebooks/03_NBA_2024_25_merge.ipynb <==
# 🏀 NBA 2024-2025: Slumps & Recoveries
## Notebook 03 - Data Merging
This notebook merges the cleaned game logs data (from Notebook 02) with:

- [NBA 2024-2025 Per Game Stats](https://www.basketball-reference.com/leagues/NBA_2025_per_game.html) (copied and pasted into `.txt` file)
- [NBA 2024-2025 Shooting Stats](https://www.basketball-reference.com/leagues/NBA_2025_shooting.html) (copied and pasted into `.txt` file)

The goal is to add **three additional columns** to the cleaned game logs data:

- position
- age
- average shot distance (season-long aggregate)

# Import libraries
import pandas as pd
import numpy as np

___
## Load Data

# Load cleaned game logs (from Notebook 02)
cleaned = pd.read_csv(r"...\03_python_outputs\01_cleaning\NBA_2024_25_game_logs.csv")

print(f"Loaded: {cleaned.shape[0]:,} rows | {cleaned["player_name"].nunique():,} players")

# Load Per Game Stats .txt file
per_game = pd.read_csv(r"...\01_original_data\to_merge\NBA_league_leaders_per_game.txt")

print(f"Loaded: {per_game.shape[0]:,} rows | {per_game["player_name"].nunique():,} players")

# Load Shooting Stats .txt file
shooting_stats = pd.read_csv(r"...\01_original_data\to_merge\NBA_league_leaders_shooting.txt")

print(f"Loaded: {shooting_stats.shape[0]:,} rows | {shooting_stats["player_name"].nunique():,} players")

___
## Merge "Per Game Stats"

# Show all columns
per_game.columns

> Only `age` and `pos` are needed.

# Keep only `age` and `pos` columns (and `player_name` as linking field)
per_game_needed_cols = per_game[["player_name", "age", "pos"]].copy()

# Rename `pos` to `position` for clarity
per_game_needed_cols = per_game_needed_cols.rename(columns={"pos": "position"})

# Merge these 2 columns back into the cleaned game logs data
cleaned_merged = cleaned.merge(per_game_needed_cols, on="player_name", how="left")

___
## Sanity Checks

# Confirm these new columns successfully merged
cleaned_merged[["player_name", "age", "position"]]

✅

# Confirm that there are zero players without a position
cleaned_merged[cleaned_merged["position"].isna()]["player_name"].unique()

✅

___
## Merge "Shooting Stats"

# Show all columns
shooting_stats.columns

> Only `distance` is needed.

# Keep only `distance` column (and `player_name` as linking field)
shooting_stats_needed_cols = shooting_stats[["player_name", "distance"]].copy()

# Merge this column back into the cleaned game logs data
merged_final = cleaned_merged.merge(shooting_stats_needed_cols, on="player_name", how="left")

___
## Sanity Checks

# Confirm that all new columns successfully merged
merged_final[["player_name", "age", "position", "distance"]]

# Confirm that there are zero players without a distance
merged_final[merged_final["distance"].isna()]["player_name"].unique()

✅

___
## Save

The data has been prepped and is now ready for analysis in the next notebook (Notebook 04).

# Save to CSV
merged_final.to_csv("NBA_2024_25_game_logs_merged.csv", index=False)
